requirements.txt

langchain-community==0.3.17
langchain==0.3.18
langgraph==0.2.72
langchain-openai==0.3.5
python-dotenv==1.0.1
streamlit==1.42.0
